+++ AlchemistCode/Cards/Rare/Sublimate.cs
@@ -1,0 +1,20 @@
+using Alchemist.AlchemistCode.Powers;
+using MegaCrit.Sts2.Core.Commands;
+using MegaCrit.Sts2.Core.Entities.Cards;
+using MegaCrit.Sts2.Core.GameActions.Multiplayer;
+
+namespace Alchemist.AlchemistCode.Cards.Rare;
+
+public class Sublimate : AlchemistCard
+{
+    public Sublimate() : base(2, CardType.Power, CardRarity.Rare, TargetType.Self)
+    {
+        WithCostUpgradeBy(-1);
+        WithTip(typeof(AntitoxinPower));
+    }
+
+    protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay play)
+    {
+        await PowerCmd.Apply<SublimatePower>(choiceContext, Owner.Creature, 1, Owner.Creature, this);
+    }
+}